Added multiple features:
When clicking 'add item' to the media watch list you are sent to the search view and can search for an item to add to the media list.

The detail view shows like it does from the other views, but this time it says 'add to media list'. Save to core data saves all watchlist items with the showcase position 0 (representing the fact theyre not in the actual top3 showcase).

The table reloads and the completion handler goes back to the showcase view as intended.

Also added functionality to check for duplicate entries in the watch list. If you try and add an item thats already in there an alert will pop up and wont let you duplicate the entry (allowing you to search for another item in the search vc)