WISHLIST IMPLEMENTATION! Firstly, I recreated the core data structure for a movie item, adding a 'alreadyWatched' attribute stored as a boolean indicating whether its already watched or in the wishlist. Next, i made a custom 'segment control' sort of thing, 2 buttons in a horizontal stack that can be toggled between in order to show either the watchlist or wishlist, each of the buttons will reload the media list collectionview with the correct list of items fetched from core data. (using the predicate linked to the previously created attribute 'alreadyWatched' Next, I edited the functioanlity of the savemovietocoredata function to successfully save the movies with the new attribute.
finally to bug fix i had to change the logic related to stopping duplicate additions, this means that there will no longer be issues about 'creating duplicates' in the list as duplicates can now be seperated by the alreadyWatched attribute.

In order to change the layout to put in the horizontal stack to switch between the watchlist/wishlist, I had to put the 'add media' button in a different place, i decided to do this by adding a first cell to each of the collection views which acts and is displayed as the + button.